General purpose string-replacement .py file
Serves as a way to replace a string in a file with a given value and optionally, truncates the string. Arguments are parsed into the .py file using the appropriate shell and if everything goes according to plan, the original string in the text file has been replaced with a new one.