Manipulate builds in depots on Steam
PREFACE: Download this package in order to use this .py script -- https://pypi.org/project/steamctl/. Additionally, append depot names on Steamworks with numbers, for example, 'depot_slot_0', 'depot_slot_1', etc.
Description: Arguments are parsed into the .py script using the appropriate shell. The procedure in the .py script then uses the JSON module to fetch the requisite JSON data from Steam, extracts the build IDs in the JSON file and performs a 'shuffle'. The shuffle is very simple -- a 'for' loop cycles through a dictionary of objects, i.e. the build IDs, increments a value to correspond with the appropriate depot name, then sends a POST request using the steamctl CLI utility to move the build down. Then repeats.